requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Flask==3.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.2.0
pytesseract==0.3.10
schedule==1.2.1 